ADEN-13924 - update after CR: - better refactor to remove code duplication, chuck code more, to continue process of size reduction - minor updates, like renaming dir, or message name, or functions
