+++ src/ad-products/templates/index.ts
@@ -3,6 +3,6 @@ export * from './interface/close-button';
 export * from './interface/video/ui-elements';
 export * from './interface/video/video-interface';
 export * from './uap';
-export * from './uap/interface/uap-config';
-export * from './uap/interface/uap-image';
-export * from './uap/interface/uap-params';
+export * from './uap/types/uap-config';
+export * from './uap/types/uap-image';
+export * from './uap/types/uap-params';
