+++ src/ad-products/templates/uap/index.ts
@@ -1,3 +1,2 @@
 export { resolvedState } from './resolved-state';
-export * from './uap-video-settings';
 export * from './universal-ad-package';
