+++ src/ad-products/templates/uap/types/uap-config.ts
