+++ src/ad-products/templates/uap/types/uap-image.ts
