+++ src/platforms/news-and-ratings/tvguide-mtc/setup/tracking/tracking.setup.ts
@@ -1,28 +1,15 @@
-import { DataWarehouseTracker, trackingUrls } from '@platforms/shared';
-import {
-	adClickTracker,
-	Apstag,
-	communicationService,
-	ctaTracker,
-	Dictionary,
-	eventsRepository,
-	GAMOrigins,
-	globalAction,
-	InstantConfigCacheStorage,
-	InstantConfigService,
-	interventionTracker,
-	porvataTracker,
-	PostmessageTracker,
-	targetingService,
-	TrackingMessage,
-	TrackingTarget,
-} from '@wikia/ad-engine';
+import { DataWarehouseTracker } from '@platforms/shared';
+import { InstantConfigService, interventionTracker } from '@wikia/ad-engine';
 import { Injectable } from '@wikia/dependency-injection';
-import { props } from 'ts-action';
 import { AdSizeTracker } from '../../../../shared/tracking/ad-size-tracker';
 import { LabradorTracker } from '../../../../shared/tracking/labrador-tracker';
-
-const adClickedAction = globalAction('[AdEngine] Ad clicked', props<Dictionary>());
+import { adClickDwTracker } from '../../../../shared/tracking/trackers/ad-click-tracker';
+import { ctaDwTracker } from '../../../../shared/tracking/trackers/cta-tracker';
+import { experimentGroupsTracker } from '../../../../shared/tracking/trackers/experiment-group-tracker';
+import { googleTopicsTracker } from '../../../../shared/tracking/trackers/google-topics-tracker';
+import { identityDwTracker } from '../../../../shared/tracking/trackers/identity-tracker';
+import { porvataDwTracker } from '../../../../shared/tracking/trackers/porvata-tracker';
+import { postmessageTrackingDwTracker } from '../../../../shared/tracking/trackers/postmessage-tracker';
 
 @Injectable()
 export class TrackingSetup {
@@ -34,132 +21,18 @@ export class TrackingSetup {
 	) {}
 
 	execute(): void {
-		this.porvataTracker();
-		this.postmessageTrackingTracker();
-		this.experimentGroupsTracker();
+		porvataDwTracker(this.dwTracker);
+		postmessageTrackingDwTracker(this.dwTracker);
+		experimentGroupsTracker(this.labradorTracker);
 		this.interventionTracker();
-		this.adClickTracker();
-		this.ctaTracker();
-		this.identityTracker();
-		this.googleTopicsTracker();
+		adClickDwTracker(this.dwTracker);
+		ctaDwTracker(this.dwTracker);
+		identityDwTracker(this.dwTracker, this.instantConfig);
+		googleTopicsTracker(this.dwTracker);
 		this.adSizeTracker.init();
 	}
 
-	private identityTracker(): void {
-		if (!this.instantConfig.get('icIdentityPartners', false)) {
-			communicationService.on(
-				eventsRepository.IDENTITY_PARTNER_DATA_OBTAINED,
-				(eventInfo) => {
-					const { partnerName, partnerIdentityId } = eventInfo.payload;
-					this.dwTracker.track(
-						{
-							partner_name: partnerName,
-							partner_identity_id: partnerIdentityId,
-						},
-						trackingUrls.IDENTITY_INFO,
-					);
-
-					if (['liveConnect', 'MediaWiki-sha256'].includes(partnerName)) {
-						const apstag = Apstag.make();
-						apstag.init().then(() => apstag.sendHEM(partnerIdentityId));
-					}
-				},
-				false,
-			);
-		}
-	}
-
-	private porvataTracker(): void {
-		communicationService.on(
-			eventsRepository.VIDEO_PLAYER_TRACKING,
-			({ eventInfo }) => {
-				this.dwTracker.track(eventInfo, trackingUrls.AD_ENG_PLAYER_INFO);
-			},
-			false,
-		);
-
-		porvataTracker.register();
-	}
-
-	// candidate to remove
-	private ctaTracker(): void {
-		ctaTracker.register(({ data }: Dictionary) => {
-			this.dwTracker.track(data, trackingUrls.AD_ENG_AD_INFO);
-		});
-	}
-
-	// candidate to remove
-	private adClickTracker(): void {
-		adClickTracker.register(({ data }: Dictionary) => {
-			// event listeners might be outside AdEngine, f.e. in the SilverSurfer interactions module
-			communicationService.dispatch(adClickedAction(data));
-			this.dwTracker.track(data, trackingUrls.AD_ENG_AD_INFO);
-		});
-	}
-
-	private postmessageTrackingTracker(): void {
-		const postmessageTracker = new PostmessageTracker(['payload', 'target']);
-
-		postmessageTracker.register<TrackingMessage>(
-			async (message) => {
-				const { target, payload } = message;
-
-				switch (target) {
-					case TrackingTarget.GoogleAnalytics: {
-						window.ga(
-							'tracker1.send',
-							'event',
-							payload.category,
-							payload.action,
-							payload.label,
-							typeof payload.value === 'number' ? payload.value.toString() : payload.value,
-						);
-						break;
-					}
-					case TrackingTarget.DataWarehouse: {
-						this.dwTracker.track(payload);
-						break;
-					}
-					default:
-						break;
-				}
-
-				return message;
-			},
-			[window.origin, ...GAMOrigins],
-		);
-	}
-
-	private experimentGroupsTracker(): void {
-		const cacheStorage = InstantConfigCacheStorage.make();
-		const experimentsGroups = [
-			...cacheStorage.getSamplingResults(),
-			...(targetingService.get('experiment_groups') || []),
-		].join(';');
-
-		if (experimentsGroups) {
-			this.labradorTracker.track(experimentsGroups);
-		}
-	}
-
-	// candidate to remove
 	private interventionTracker(): void {
 		interventionTracker.register();
-	}
-
-	private async googleTopicsTracker(): Promise<void> {
-		if (targetingService.get('topics_available') !== '1') {
-			return;
-		}
-
-		// @ts-expect-error Google Topics API is not available in TS dom lib
-		const topics: unknown[] = await document.browsingTopics({ skipObservation: true });
-		this.dwTracker.track(
-			{
-				ppid: targetingService.get('ppid'),
-				topic: JSON.stringify(topics),
-			},
-			trackingUrls.TOPICS,
-		);
 	}
 }
