+++ src/platforms/shared/tracking/trackers/experiment-group-tracker.ts
@@ -1,0 +1,14 @@
+import { InstantConfigCacheStorage, targetingService } from '@wikia/ad-engine';
+import { LabradorTracker } from '../labrador-tracker';
+
+export function experimentGroupsTracker(labradorTracker: LabradorTracker): void {
+	const cacheStorage = InstantConfigCacheStorage.make();
+	const experimentsGroups = [
+		...cacheStorage.getSamplingResults(),
+		...(targetingService.get('experiment_groups') || []),
+	].join(';');
+
+	if (experimentsGroups) {
+		labradorTracker.track(experimentsGroups);
+	}
+}
